+++ src/Light.h
@@ -6,14 +6,14 @@
 class Light {
     public:
         int type; // 0: ambient; 1: point; 2: directional
-        float intensity;
+        double intensity;
 
         Vector3 position;
         Vector3 direction;
 
         Light();
-        Light(int t, float i);
-        Light(int t, float i, Vector3 pos_or_rot);
+        Light(int t, double i);
+        Light(int t, double i, Vector3 pos_or_rot);
 };
 
 #endif
